Created main menu and options menu
Created the Main Menu and Options menu. The main menu has the 3 buttons from the Trello board. The play will take us to the next scene. Options will take you to a separate menu with a sound toggle (does not work yet). Quit will quit the game. TODO > Update background to match the game. TODO > When sound is put in game, create toggle to interact with music of the game to mute/unmute the game.